apps/backend/requirements.txt
uvicorn==0.35.0
python-dotenv==1.0.0
langgraph==0.3.19
pydantic-ai[anthropic]==0.0.43
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1
instructor==1.7.7
langsmith
langchain
langchain_anthropic
fastapi
beautifulsoup4==4.12.2
groq==0.15.0
google-genai
jsonref
pillow>=10.0.0
python-pptx>=0.6.23
ipython==8.18.1
diskcache
arize-phoenix-otel
supabase>=2.3.5
google-search-results==2.4.2
aiohttp>=3.8.0
asyncpg>=0.29.0
openai>=1.0.0
numpy>=1.24.0
requests
psutil
email-validator>=2.2.0
# Note: json_schema_to_pydantic - using git version for Render compatibility
git+https://github.com/brandonmoak/json-schema-to-pydantic.git@main
# json-schema-to-pydantic>=0.2.0
watchfiles==1.1.0
httpx==0.27.0

# Web crawling/search
firecrawl-py

# Monitoring and Error Tracking
sentry-sdk[fastapi]==2.33.0

# Authentication Dependencies
google-auth>=2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
PyJWT==2.10.1

# Data file parsing for attachments
openpyxl>=3.1.3

# Email Service (optional - for production magic links)
resend==2.11.0  # For professional email sending
# sendgrid==6.11.0  # Uncomment if using SendGrid
# boto3==1.34.0     # Uncomment if using AWS SES

